pre-verify class files before adding to repository
rather than loading each class file along with its dependencies
in separate class loaders only to make sure it's IBasicRobot
